# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Plotted API",
    "description" : "Platform-neutral streaming decision and subscription optimisation platform. Title metadata and regional availability originate from TMDB and JustWatch; see /legal/data-sources for attribution and refresh cadence.",
    "license" : {
      "name" : "AGPL-3.0-or-later"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "This deployment"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
    "/api/v1/watchlist/items" : {
      "post" : {
        "tags" : [ "watchlist-controller" ],
        "summary" : "Add a title to the watchlist",
        "description" : "Idempotent: adding a title that is already on the list returns the existing entry rather than failing, because adding something twice is a slip rather than an error.",
        "operationId" : "add",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddWatchlistItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WatchlistItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/subscriptions" : {
      "get" : {
        "tags" : [ "subscription-controller" ],
        "summary" : "What the signed-in user pays for",
        "description" : "monthlyTotal counts active and trial subscriptions only. Cancelled and lapsed rows are kept and returned, because what someone used to pay for is part of the record, but they cost nothing now and including them would overstate the bill.",
        "operationId" : "list_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubscriptionListResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "subscription-controller" ],
        "summary" : "Record a subscription",
        "operationId" : "add_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/watchlist/items/{itemId}" : {
      "delete" : {
        "tags" : [ "watchlist-controller" ],
        "summary" : "Remove a title from the watchlist",
        "operationId" : "remove",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "watchlist-controller" ],
        "summary" : "Update priority, status or intent for a watchlist entry",
        "description" : "Partial update; omitted fields are unchanged.",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateWatchlistItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WatchlistItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/subscriptions/{subscriptionId}" : {
      "delete" : {
        "tags" : [ "subscription-controller" ],
        "summary" : "Forget a subscription entirely",
        "description" : "Deletes the record. To mark something as no longer paid for while keeping its history, PATCH the status to cancelled instead.",
        "operationId" : "remove_1",
        "parameters" : [ {
          "name" : "subscriptionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "subscription-controller" ],
        "summary" : "Update a subscription",
        "description" : "Partial update; omitted fields are unchanged.",
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "subscriptionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateSubscriptionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AddWatchlistItemRequest" : {
        "required" : [ "titleId" ],
        "type" : "object",
        "properties" : {
          "titleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "priority" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer",
            "description" : "1 is the highest priority and 5 the lowest. Defaults to 3.",
            "format" : "int32"
          },
          "desiredByDate" : {
            "type" : "string",
            "format" : "date"
          },
          "notes" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "WatchlistItemResponse" : {
        "required" : [ "addedAt", "id", "priority", "status", "titleId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "titleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "priority" : {
            "type" : "integer",
            "description" : "1 is the highest priority and 5 the lowest.",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          },
          "addedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "desiredByDate" : {
            "type" : "string",
            "format" : "date"
          },
          "notes" : {
            "type" : "string"
          },
          "title" : {
            "$ref" : "#/components/schemas/WatchlistTitleResponse"
          }
        },
        "description" : "A title on the signed-in user's watchlist."
      },
      "WatchlistTitleResponse" : {
        "required" : [ "mediaType", "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "mediaType" : {
            "type" : "string"
          },
          "releaseYear" : {
            "type" : "integer",
            "format" : "int32"
          },
          "posterUrl" : {
            "type" : "string"
          },
          "watchMinutes" : {
            "type" : "integer",
            "description" : "Runtime for a film, total episode runtime for a series. Null when unknown.",
            "format" : "int32"
          }
        },
        "description" : "Null when the title has been removed from the catalogue since it was added."
      },
...
      "CreateSubscriptionRequest" : {
        "required" : [ "autoRenews", "cannotCancel", "price", "providerId" ],
        "type" : "object",
        "properties" : {
          "providerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "planName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string",
            "description" : "The plan as the user knows it, for example 'Standard with ads'. Defaults to 'Standard'."
          },
          "billingPeriod" : {
            "type" : "string",
            "description" : "One of monthly, quarterly, annual.",
            "enum" : [ "monthly", "quarterly", "annual" ]
          },
          "price" : {
            "minimum" : 0.0,
            "exclusiveMinimum" : false,
            "type" : "number"
          },
          "currency" : {
            "maxLength" : 3,
            "minLength" : 3,
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "description" : "One of active, paused, cancelled, trial, lapsed. Defaults to active.",
            "enum" : [ "active", "paused", "cancelled", "trial", "lapsed" ]
          },
          "startedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "renewsOn" : {
            "type" : "string",
            "format" : "date"
          },
          "commitmentEndsOn" : {
            "type" : "string",
            "format" : "date"
          },
          "autoRenews" : {
            "type" : "boolean"
          },
          "cannotCancel" : {
            "type" : "boolean"
          },
          "notes" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "description" : "Records a subscription the user pays for. Price is supplied by the user rather than looked up: Plotted deliberately ships no pricing data, because a figure it invented would flow straight into the cancellation optimiser's objective."
      },
      "SubscriptionResponse" : {
        "required" : [ "autoRenews", "billingPeriod", "cannotCancel", "currency", "id", "monthlyCost", "planName", "price", "providerId", "providerName", "providerSlug", "startedOn", "status" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "providerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "providerName" : {
            "type" : "string"
          },
          "providerSlug" : {
            "type" : "string"
          },
          "providerLogoUrl" : {
            "type" : "string"
          },
          "planName" : {
            "type" : "string"
          },
          "billingPeriod" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number",
            "description" : "What the user told us they pay, per billing period."
          },
          "monthlyCost" : {
            "type" : "number",
            "description" : "The same figure expressed per month, so plans on different cycles compare."
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "startedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "renewsOn" : {
            "type" : "string",
            "format" : "date"
          },
          "autoRenews" : {
            "type" : "boolean"
          },
          "cannotCancel" : {
            "type" : "boolean",
            "description" : "Inside a commitment. Phase 5 treats this as a hard constraint, never a cost."
          },
          "commitmentEndsOn" : {
            "type" : "string",
            "format" : "date"
          },
          "notes" : {
            "type" : "string"
          }
        }
      },
...
      "UpdateWatchlistItemRequest" : {
        "required" : [ "clearDesiredByDate", "clearNotes" ],
        "type" : "object",
        "properties" : {
          "priority" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "description" : "One of pending, in_progress, completed, abandoned, unavailable.",
            "enum" : [ "pending", "in_progress", "completed", "abandoned", "unavailable" ]
          },
          "desiredByDate" : {
            "type" : "string",
            "format" : "date"
          },
          "clearDesiredByDate" : {
            "type" : "boolean",
            "description" : "Set true to remove an existing desiredByDate; JSON null alone cannot express this."
          },
          "notes" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "clearNotes" : {
            "type" : "boolean",
            "description" : "Set true to remove existing notes."
          }
        }
      },
...
      "UpdateSubscriptionRequest" : {
        "required" : [ "clearNotes", "clearRenewsOn" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "active", "paused", "cancelled", "trial", "lapsed" ]
          },
          "renewsOn" : {
            "type" : "string",
            "format" : "date"
          },
          "clearRenewsOn" : {
            "type" : "boolean",
            "description" : "Set true to remove an existing renewal date."
          },
          "autoRenews" : {
            "type" : "boolean"
          },
          "cannotCancel" : {
            "type" : "boolean"
          },
          "notes" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "clearNotes" : {
            "type" : "boolean",
            "description" : "Set true to remove existing notes."
          }
        }
      },
...
      "SubscriptionListResponse" : {
        "required" : [ "countedSubscriptions", "currency", "monthlyTotal", "subscriptions" ],
        "type" : "object",
        "properties" : {
          "subscriptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubscriptionResponse"
            }
          },
          "monthlyTotal" : {
            "type" : "number",
            "description" : "Total monthly cost of active and trial subscriptions only."
          },
          "currency" : {
            "type" : "string"
          },
          "countedSubscriptions" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Short-lived access token issued by /api/v1/auth/login.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}